Extract columnTypeToClass into ColumnTypeConverter to reuse on client and server
