+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/Commons.java
@@ -703,74 +703,6 @@ public final class Commons {
     }
 
     /**
-     * Column type to Java class.
-     */
-    public static Class<?> columnTypeToClass(SqlColumnType type) {
-        assert type != null;
-
-        switch (type) {
-            case BOOLEAN:
-                return Boolean.class;
-            case INT8:
-                return Byte.class;
-
-            case INT16:
-                return Short.class;
-
-            case INT32:
-                return Integer.class;
-
-            case INT64:
-                return Long.class;
-
-            case FLOAT:
-                return Float.class;
-
-            case DOUBLE:
-                return Double.class;
-
-            case NUMBER:
-                return BigInteger.class;
-
-            case DECIMAL:
-                return BigDecimal.class;
-
-            case UUID:
-                return UUID.class;
-
-            case STRING:
-                return String.class;
-
-            case BYTE_ARRAY:
-                return byte[].class;
-
-            case BITMASK:
-                return BitSet.class;
-
-            case DATE:
-                return LocalDate.class;
-
-            case TIME:
-                return LocalTime.class;
-
-            case DATETIME:
-                return LocalDateTime.class;
-
-            case TIMESTAMP:
-                return Instant.class;
-
-            case PERIOD:
-                return Period.class;
-
-            case DURATION:
-                return Duration.class;
-
-            default:
-                throw new IllegalArgumentException("Unsupported type " + type);
-        }
-    }
-
-    /**
      * NativeTypePrecision.
      * TODO Documentation https://issues.apache.org/jira/browse/IGNITE-15859
      */
